+++ OpenTabletDriver.Plugin/Tablet/IRotationReport.cs
@@ -2,6 +2,6 @@ namespace OpenTabletDriver.Plugin.Tablet
 {
     public interface IRotationReport : IDeviceReport
     {
-        double Rotation { set; get; }
+        uint Rotation { set; get; }
     }
 }
